requirements.txt
flask
flask-cors
pillow
torch
diffusers
transformers
accelerate